Afrapportering/requirements.txt
certifi==2025.4.26
cffi==1.17.1
chardet==5.2.0
charset-normalizer==3.4.2
contourpy==1.3.2
cycler==0.12.1
et_xmlfile==2.0.0
fonttools==4.58.1
holidays==0.73
idna==3.10
kiwisolver==1.4.8
Markdown==3.8
matplotlib==3.10.3
numpy==2.2.6
openpyxl==3.1.5
orjson==3.10.18
packaging==25.0
pandas==2.2.3
pillow==11.2.1
pycparser==2.22
pyparsing==3.2.3
PyPDF2==3.0.1
PySide6==6.9.0
PySide6_Addons==6.9.0
PySide6_Essentials==6.9.0
python-dateutil==2.9.0.post0
pytz==2025.2
reportlab==4.4.1
requests==2.32.3
shiboken6==6.9.0
six==1.17.0
sounddevice==0.5.2
tzdata==2025.2
urllib3==2.4.0
